pyproject.toml
[tool.poetry]
name = "rexmapdb"
version = "0.1.0"
description = ""
authors = ["igor-sb <84479009+igor-sb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycurl = "^7.45.2"
biopython = "^1.81"
numpy = "^1.25.1"
pandas = "^2.0.3"
six = "^1.16.0"
fire = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
